ci: the map view runs on every pull request, and eleven findings behind it
The v0.1.0 tag settled the argument for this. A field was added to
SeasonChart::Line and one of its three callers built a Line with an
aggregate initialiser nobody updated. MSVC says nothing about a missing
initialiser; gcc and clang stop the build. Nothing else compiles the GUI
with gcc or clang, so the first thing to try it was the release: it
failed to package on Linux and macOS, and the tag pointed at a commit
that could not be built.

The workflow already documented the gap and asked a person to run the job
from the Actions tab whenever the GUI changed. A step someone has to
remember is not a gate, and this one was written down and still missed.
Eight minutes a pull request is cheaper than finding it at a tag.

Running it turned up eleven clang-tidy findings that had accumulated
behind the same gap, and they are fixed here so the job can be blocking
from its first run rather than red on arrival:

  narrowing qsizetype to int      TrendDialog x2, main.cpp
  nested conditional operators    SeasonChart, SetupPanel
  QPen that can be const          SeasonChart
  C-style array and its name      DashboardDialog
  return type repeated            DashboardDialog x3

Three more came out of running clang-tidy locally against the desktop
build, which CI's Linux job does not report: an int64 byte count divided
by a multiplication performed in int. std::int64_t is long long on
Windows and long on Linux, so which compiler notices depends on the
platform - the same trap CLAUDE.md warns about, in the other direction.

app/src/gui and viz/ now lint clean under Windows clang-tidy, which is
the stricter of the two.

+++ app/src/gui/main.cpp
@@ -399,7 +399,7 @@ int main(int argc, char** argv) {
             QCoreApplication::processEvents();
             QString each = path;
             if (tabs != nullptr && pages > 1) {
-              const int dot = each.lastIndexOf('.');
+              const int dot = static_cast<int>(each.lastIndexOf('.'));
               const QString stem = dot > 0 ? each.left(dot) : each;
               const QString suffix = dot > 0 ? each.mid(dot) : QString(".png");
               each = stem + "-" + tabs->tabText(page).toLower().replace(' ', '-') + suffix;
